requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
whitenoise==6.7.0
gunicorn==23.0.0
requests==2.32.3
django-cors-headers==4.4.0

# Testing
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==7.0.0

# Code Quality
flake8==7.3.0
black==25.9.0
isort==7.0.0
autoflake==2.3.1
radon==6.0.1